+++ packages/voter-stake-registry-hooks/src/hooks/useClaimAllPositionsRewards.ts
@@ -1,5 +1,4 @@
 import { BN, Program } from "@coral-xyz/anchor";
-import { useSolanaUnixNow } from "@helium/helium-react-hooks";
 import {
   EPOCH_LENGTH,
   PROGRAM_ID,
@@ -11,16 +10,17 @@ import { PublicKey, TransactionInstruction } from "@solana/web3.js";
 import { useAsyncCallback } from "react-async-hook";
 import { useHeliumVsrState } from "../contexts/heliumVsrContext";
 import { PositionWithMeta } from "../sdk/types";
+import { MAX_TRANSACTIONS_PER_SIGNATURE_BATCH } from "../constants";
 
 export const useClaimAllPositionsRewards = () => {
-  const { provider } = useHeliumVsrState();
-  const unixNow = useSolanaUnixNow();
+  const { provider, unixNow } = useHeliumVsrState();
   const { error, loading, execute } = useAsyncCallback(
     async ({
       positions,
       programId = PROGRAM_ID,
       onProgress,
       onInstructions,
+      maxSignatureBatch = MAX_TRANSACTIONS_PER_SIGNATURE_BATCH,
     }: {
       positions: PositionWithMeta[];
       programId?: PublicKey;
@@ -29,6 +29,7 @@ export const useClaimAllPositionsRewards = () => {
       onInstructions?: (
         instructions: TransactionInstruction[]
       ) => Promise<void>;
+      maxSignatureBatch?: number;
     }) => {
       const isInvalid =
         !unixNow || !provider || !positions.every((pos) => pos.hasRewards);
@@ -81,7 +82,10 @@ export const useClaimAllPositionsRewards = () => {
           await batchParallelInstructions(
             provider,
             multiDemArray.flat(),
-            onProgress
+            onProgress,
+            10,
+            [],
+            maxSignatureBatch
           );
         }
       }