+++ packages/voter-stake-registry-hooks/src/hooks/useClaimPositionRewards.ts
@@ -1,26 +1,27 @@
 import { BN, Program } from "@coral-xyz/anchor";
-import { useSolanaUnixNow } from "@helium/helium-react-hooks";
 import {
   EPOCH_LENGTH,
   PROGRAM_ID,
   delegatedPositionKey,
   init,
 } from "@helium/helium-sub-daos-sdk";
-import { batchParallelInstructions, Status } from "@helium/spl-utils";
+import { Status, batchParallelInstructions } from "@helium/spl-utils";
+import { isClaimed } from "@helium/voter-stake-registry-sdk";
 import { PublicKey, TransactionInstruction } from "@solana/web3.js";
 import { useAsyncCallback } from "react-async-hook";
+import { MAX_TRANSACTIONS_PER_SIGNATURE_BATCH } from "../constants";
 import { useHeliumVsrState } from "../contexts/heliumVsrContext";
 import { PositionWithMeta } from "../sdk/types";
 
 export const useClaimPositionRewards = () => {
-  const { provider } = useHeliumVsrState();
-  const unixNow = useSolanaUnixNow();
+  const { provider, unixNow } = useHeliumVsrState();
   const { error, loading, execute } = useAsyncCallback(
     async ({
       position,
       programId = PROGRAM_ID,
       onProgress,
       onInstructions,
+      maxSignatureBatch = MAX_TRANSACTIONS_PER_SIGNATURE_BATCH,
     }: {
       position: PositionWithMeta;
       programId?: PublicKey;
@@ -29,6 +30,7 @@ export const useClaimPositionRewards = () => {
       onInstructions?: (
         instructions: TransactionInstruction[]
       ) => Promise<void>;
+      maxSignatureBatch?: number;
     }) => {
       const isInvalid = !unixNow || !provider || !position.hasRewards;
 
@@ -45,11 +47,18 @@ export const useClaimPositionRewards = () => {
         const delegatedPosAcc =
           await hsdProgram.account.delegatedPositionV0.fetch(delegatedPosKey);
 
-        const { lastClaimedEpoch } = delegatedPosAcc;
+        const { lastClaimedEpoch, claimedEpochsBitmap } = delegatedPosAcc;
         const epoch = lastClaimedEpoch.add(new BN(1));
         const epochsToClaim = Array.from(
           { length: currentEpoch.sub(epoch).toNumber() },
           (_v, k) => epoch.addn(k)
+        ).filter(
+          (epoch) =>
+            !isClaimed({
+              epoch: epoch.toNumber(),
+              lastClaimedEpoch: lastClaimedEpoch.toNumber(),
+              claimedEpochsBitmap,
+            })
         );
 
         const instructions: TransactionInstruction[] = await Promise.all(
@@ -70,7 +79,14 @@ export const useClaimPositionRewards = () => {
         if (onInstructions) {
           await onInstructions(instructions);
         } else {
-          await batchParallelInstructions(provider, instructions, onProgress);
+          await batchParallelInstructions(
+            provider,
+            instructions,
+            onProgress,
+            10,
+            [],
+            maxSignatureBatch
+          );
         }
       }
     }