+++ programs/hexboosting/src/instructions/mod.rs
@@ -2,8 +2,10 @@ pub mod boost_v0;
 pub mod close_boost_v0;
 pub mod initialize_boost_config_v0;
 pub mod start_boost_v0;
+pub mod update_boost_config_v0;
 
 pub use boost_v0::*;
 pub use close_boost_v0::*;
 pub use initialize_boost_config_v0::*;
 pub use start_boost_v0::*;
+pub use update_boost_config_v0::*;