+++ packages/account-postgres-sink-service/CHANGELOG.md
@@ -3,6 +3,68 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.27](https://github.com/helium/helium-program-libary/compare/v0.6.25...v0.6.27) (2024-02-01)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
+## [0.6.26](https://github.com/helium/helium-program-libary/compare/v0.6.5...v0.6.26) (2024-01-31)
+
+
+### Features
+
+* **#543:** Add transaction webhook capability to account-postgres-sink ([#544](https://github.com/helium/helium-program-libary/issues/544)) ([a0d7bd9](https://github.com/helium/helium-program-libary/commit/a0d7bd9cbf6e5a28c467d8df67c9702cef705dae)), closes [#543](https://github.com/helium/helium-program-libary/issues/543)
+
+
+
+
+
+## [0.6.25](https://github.com/helium/helium-program-libary/compare/v0.6.24...v0.6.25) (2024-01-29)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
+## [0.6.24](https://github.com/helium/helium-program-libary/compare/v0.6.23...v0.6.24) (2024-01-29)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
+## [0.6.23](https://github.com/helium/helium-program-libary/compare/v0.6.20...v0.6.23) (2024-01-20)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
+## [0.6.22](https://github.com/helium/helium-program-libary/compare/v0.6.5...v0.6.22) (2024-01-19)
+
+
+### Features
+
+* **#543:** Add transaction webhook capability to account-postgres-sink ([#544](https://github.com/helium/helium-program-libary/issues/544)) ([a0d7bd9](https://github.com/helium/helium-program-libary/commit/a0d7bd9cbf6e5a28c467d8df67c9702cef705dae)), closes [#543](https://github.com/helium/helium-program-libary/issues/543)
+
+
+
+
+
+## [0.6.21](https://github.com/helium/helium-program-libary/compare/v0.6.20...v0.6.21) (2024-01-19)
+
+**Note:** Version bump only for package @helium/account-postgres-sink-service
+
+
+
+
+
 ## [0.6.20](https://github.com/helium/helium-program-libary/compare/v0.6.19...v0.6.20) (2024-01-19)
 
 **Note:** Version bump only for package @helium/account-postgres-sink-service