+++ packages/spl-utils/CHANGELOG.md
@@ -3,6 +3,68 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.27](https://github.com/helium/helium-program-library/compare/v0.6.25...v0.6.27) (2024-02-01)
+
+**Note:** Version bump only for package @helium/spl-utils
+
+
+
+
+
+## [0.6.26](https://github.com/helium/helium-program-library/compare/v0.6.5...v0.6.26) (2024-01-31)
+
+
+### Features
+
+* **#504:** Parallelize claim delegation rewards  ([#509](https://github.com/helium/helium-program-library/issues/509)) ([01b5f3d](https://github.com/helium/helium-program-library/commit/01b5f3dbd35345302e5fc6a83667b5e5e017d4ec)), closes [#504](https://github.com/helium/helium-program-library/issues/504)
+
+
+
+
+
+## [0.6.25](https://github.com/helium/helium-program-library/compare/v0.6.24...v0.6.25) (2024-01-29)
+
+**Note:** Version bump only for package @helium/spl-utils
+
+
+
+
+
+## [0.6.24](https://github.com/helium/helium-program-library/compare/v0.6.23...v0.6.24) (2024-01-29)
+
+**Note:** Version bump only for package @helium/spl-utils
+
+
+
+
+
+## [0.6.23](https://github.com/helium/helium-program-library/compare/v0.6.20...v0.6.23) (2024-01-20)
+
+**Note:** Version bump only for package @helium/spl-utils
+
+
+
+
+
+## [0.6.22](https://github.com/helium/helium-program-library/compare/v0.6.5...v0.6.22) (2024-01-19)
+
+
+### Features
+
+* **#504:** Parallelize claim delegation rewards  ([#509](https://github.com/helium/helium-program-library/issues/509)) ([01b5f3d](https://github.com/helium/helium-program-library/commit/01b5f3dbd35345302e5fc6a83667b5e5e017d4ec)), closes [#504](https://github.com/helium/helium-program-library/issues/504)
+
+
+
+
+
+## [0.6.21](https://github.com/helium/helium-program-library/compare/v0.6.20...v0.6.21) (2024-01-19)
+
+**Note:** Version bump only for package @helium/spl-utils
+
+
+
+
+
 ## [0.6.20](https://github.com/helium/helium-program-library/compare/v0.6.19...v0.6.20) (2024-01-19)
 
 **Note:** Version bump only for package @helium/spl-utils